requirements.txt
flask==2.2.3
flask-apscheduler==1.12.4
flask-pymongo==2.3.0
requests==2.28.2
python-dateutil==2.8.2